+++ guns-main/src/main/java/cn/stylefeng/guns/modular/scscore/controller/ScScoreController.java
@@ -84,7 +84,7 @@ public class ScScoreController {
     @Permission
     @GetMapping("/scScore/export/{id}")
     @BusinessLog(title = "考试成绩_导出", opType = LogAnnotionOpTypeEnum.EXPORT)
-    public void export(@PathVariable(value = "id", required = false) Integer[] id,
+    public void export(@PathVariable(value = "id", required = false) Long[] id,
                        ScAdmissionTypeEnum scAdmissionTypeEnum) {
         scScoreService.export(id,scAdmissionTypeEnum);
     }
